services/user-auth-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
pydantic[email]==2.5.0
python-multipart==0.0.6
PyJWT==2.8.0
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
alembic==1.13.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
slowapi==0.1.9
redis==5.0.1
aioredis==2.0.1